refactor: atom 선언과 Recoil함수 선언을 통합
